requirements.txt
certifi==2020.6.20
charset-normalizer==2.0.12
idna==3.3
PyYAML==6.0
requests==2.27.1
urllib3==1.26.9
